requirements.txt
Django==5.1.6
djangorestframework==3.15.2
pillow==11.1.0
Faker==36.11